pyproject.toml
[project]
name = "nonebot-plugin-number-detection"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "zhongwen-4", email = "2401128923@qq.com"},
]
dependencies = [
    "nonebot2>=2.4.1",
    "nonebot-adapter-onebot>=2.4.6",
]
requires-python = ">=3.9.0"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
